requirements.txt
spotipy==2.24.0
python-dotenv==1.0.1
pandas==2.2.3
plotly==5.24.1
streamlit==1.39.0
numpy==2.1.3

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
coverage>=7.3.0
